requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.2
pytz==2021.1
reportlab==3.6.1
sqlparse==0.4.1
